+++ lib/cocina_display/description/access_contact.rb
@@ -1,5 +1,5 @@
 module CocinaDisplay
-  module Accesses
+  module Description
     class AccessContact < Access
       # Whether the access contact info is a contact email.
       # @return [Boolean]
